feat: mypage에 뷰모델 적용
